Make changes for the "Import Document" button
